HSEARCH-4498 Remove unnecessary "Search" keyword from the name of projection IT classes
